Move dims models to components directory
