--- napari/components/_dims/__init__.py
@@ -1,1 +1,0 @@
-from .dims import Dims
