+++ napari/components/dims_constants.py
